+++ jdaviz/configs/cubeviz/plugins/viewers.py
@@ -8,6 +8,15 @@ __all__ = ['CubevizImageView']
 
 @viewer_registry("cubeviz-image-viewer", label="Image 2D (Cubeviz)")
 class CubevizImageView(BqplotImageView):
+    # Whether to inherit tools from glue-jupyter automatically. Set this to
+    # False to have full control here over which tools are shown in case new
+    # ones are added in glue-jupyter in future that we don't want here.
+    inherit_tools = False
+
+    tools = ['bqplot:home', 'jdaviz:boxzoom',
+             'bqplot:panzoom', 'bqplot:rectangle',
+             'bqplot:circle']
+
     default_class = None
 
     def set_plot_axes(self):