+++ jdaviz/configs/imviz/plugins/viewers.py
@@ -19,8 +19,9 @@ class ImvizImageView(BqplotImageView, AstrowidgetsImageViewerMixin):
     # ones are added in glue-jupyter in future that we don't want here.
     inherit_tools = False
 
-    tools = ['bqplot:home', 'bqplot:panzoom', 'bqplot:panzoommatch',
-             'bqplot:contrastbias', 'bqplot:blinkonce',
+    tools = ['bqplot:home', 'jdaviz:boxzoom', 'jdaviz:boxzoommatch',
+             'bqplot:panzoom', 'jdaviz:panzoommatch',
+             'jdaviz:contrastbias', 'jdaviz:blinkonce',
              'bqplot:rectangle', 'bqplot:circle', 'bqplot:ellipse']
     default_class = None